bug: 21744 Fix rounding in displaySizePercent, add '%' to lite client display
http://bugzilla.zimbra.com/show_bug.cgi?id=21744

+++ ZimbraTagLib/src/java/com/zimbra/cs/taglib/bean/BeanUtils.java
@@ -301,7 +301,7 @@ public class BeanUtils {
     public static String displaySizePercent(long size, long max) {
         String formt = " ";
         double dsize;
-        dsize = (size*100)/max;
+        dsize = (size*100)/(double)max;
         dsize = Math.round(dsize);
         return dsize + formt;
     }
